+++ qupath-extension-script-editor/src/main/java/qupath/lib/gui/scripting/richtextfx/stylers/GroovyStyler.java
@@ -2,9 +2,7 @@
  * #%L
  * This file is part of QuPath.
  * %%
- * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
- * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2022 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2022 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -23,30 +21,30 @@
 
 package qupath.lib.gui.scripting.richtextfx.stylers;
 
+import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
+import java.util.HashSet;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
+import org.apache.commons.lang3.math.NumberUtils;
 import org.fxmisc.richtext.CodeArea;
 import org.fxmisc.richtext.model.StyleSpans;
-import org.fxmisc.richtext.model.StyleSpansBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import qupath.lib.common.LogTools;
-
 /**
  * Styling to apply to a {@link CodeArea}, based on Groovy syntax.
- * @author Melvin Gelbard
- * @since v0.4.0
+ * @author Pete Bankhead
+ * @since v0.4.1
+ * 
+ * @implNote This was rewritten for v0.4.1 to avoid relying on regex, 
+ *           which could sometimes result in a {@link StackOverflowError}.
  */
 public class GroovyStyler implements ScriptStyler {
 	
 	private static final Logger logger = LoggerFactory.getLogger(GroovyStyler.class);
 	
-	private static final String[] KEYWORDS = new String[] {
+	private static final Set<String> KEYWORDS = new HashSet<>(Arrays.asList(
             "abstract", "assert", "boolean", "break", "byte",
             "case", "catch", "char", "class", "const",
             "continue", "default", "do", "double", "else",
@@ -58,36 +56,7 @@ public class GroovyStyler implements ScriptStyler {
             "switch", "synchronized", "this", "throw", "throws",
             "transient", "try", "void", "volatile", "while",
             "def", "in", "with", "trait", "true", "false", "var"
-    };
-	
-	private static Pattern PATTERN;
-	
-	static {
-		final String KEYWORD_PATTERN = "\\b(" + String.join("|", KEYWORDS) + ")\\b";
-//		final String METHOD_PATTERN = "[a-zA-Z]+\\(";
-//		final String METHOD_PATTERN = "\\b(" + String.join("|", METHOD_NAMES) + ")\\b";
-	    final String PAREN_PATTERN = "\\(|\\)";
-	    final String BRACE_PATTERN = "\\{|\\}";
-	    final String BRACKET_PATTERN = "\\[|\\]";
-	    final String SEMICOLON_PATTERN = "\\;";
-	    final String TRIPLE_QUOTE_PATTERN = "\"\"\"([^\"\"\"\\\\]|\\\\.)*\"\"\"";
-	    final String DOUBLE_QUOTE_PATTERN = "\"([^\"\\\\]|\\\\.)*\"";
-	    final String SINGLE_QUOTE_PATTERN = "'([^'\\\\]|\\\\.)*\'";
-	    final String COMMENT_PATTERN = "//[^\n]*" + "|" + "/\\*(.|\\R)*?\\*/";
-	    
-	    PATTERN = Pattern.compile(
-	            "(?<KEYWORD>" + KEYWORD_PATTERN + ")"
-	            + "|(?<PAREN>" + PAREN_PATTERN + ")"
-	            + "|(?<BRACE>" + BRACE_PATTERN + ")"
-	            + "|(?<BRACKET>" + BRACKET_PATTERN + ")"
-	            + "|(?<SEMICOLON>" + SEMICOLON_PATTERN + ")"
-	            + "|(?<TRIPLEQUOTES>" + TRIPLE_QUOTE_PATTERN + ")"
-	            + "|(?<DOUBLEQUOTES>" + DOUBLE_QUOTE_PATTERN + ")"
-	            + "|(?<SINGLEQUOTES>" + SINGLE_QUOTE_PATTERN + ")"
-	            + "|(?<COMMENT>" + COMMENT_PATTERN + ")"
-	    );
-	    
-	}
+    ));
 	
 	/**
 	 * Constructor.
@@ -99,39 +68,173 @@ public class GroovyStyler implements ScriptStyler {
 		return Set.of("groovy", "java");
 	}
 	
+	
 	@Override
 	public StyleSpans<Collection<String>> computeEditorStyles(final String text) {
 		
-		// Stop quickly if we have really long lines, which can cause QuPath to freeze
-		int longLine = ScriptStylerTools.DEFAULT_LONG_LINE_LENGTH;
-		if (ScriptStylerTools.containsLongLines(text, longLine)) {
-			LogTools.logOnce(logger, "Text contains lines longer than " + longLine + " - no styling will be applied");
-			return ScriptStylerProvider.getPlainStyling(text);
+		long startTime = System.currentTimeMillis();
+		
+		var visitor = new StyleSpanVisitor(text);
+		
+		var chars = text.toCharArray();
+		int ind = 0;
+		int n = chars.length;
+		var buffer = new StringBuffer();
+		while (ind < n) {
+			// Get current and previous character
+			char c = chars[ind];
+			char lastChar = ind > 0 ? chars[ind-1] : '\n';
+			
+			switch (c) {
+			case '(':
+			case ')':
+				handleToken(visitor, ind, buffer);
+				handleSingleCharacterStyle(visitor, ind, "paren");
+				break;
+			case '[':
+			case ']':
+				handleToken(visitor, ind, buffer);
+				handleSingleCharacterStyle(visitor, ind, "bracket");
+				break;
+			case '{':
+			case '}':
+				handleToken(visitor, ind, buffer);
+				handleSingleCharacterStyle(visitor, ind, "brace");
+				break;
+			case ';':
+				handleToken(visitor, ind, buffer);
+				handleSingleCharacterStyle(visitor, ind, "semicolon");
+				break;
+			case '\\':
+				if (lastChar == '\\') {
+					// Handle line comment
+					handleToken(visitor, ind-1, buffer);
+					visitor.appendStyle(ind-1);
+					visitor.push("comment");
+					while (ind < n && chars[ind] != '\n')
+						ind++;
+				}
+				break;
+			case ',':
+				handleToken(visitor, ind, buffer);
+				break;
+			case '=':
+				handleToken(visitor, ind, buffer);
+				break;
+			case '*':
+				if (lastChar == '/') {
+					// Handle block comment
+					handleToken(visitor, ind-1, buffer);
+					visitor.appendStyle(ind-1);
+					visitor.push("comment");
+					ind++;
+					while (ind < n && !(chars[ind] == '/' && chars[ind-1] == '*')) {
+						ind++;
+					}
+					visitor.appendStyle(ind);
+					visitor.pop();
+				}
+				break;
+			case '\'':
+				// Handle single quotes
+				visitor.appendStyle(ind);
+				visitor.push("string");
+				ind++;
+				while (ind < n && chars[ind] != '\'' && chars[ind] != '\n')
+					ind++;
+				visitor.appendStyle(ind+1);
+				visitor.pop();				
+				break;
+			case '"':
+				// Handle double or triple quotes
+				resetToken(buffer);
+				visitor.appendStyle(ind);
+				visitor.push("string");
+				if (ind < n - 2 && chars[ind+1] == '"' && chars[ind+2] == '"') {
+					// Handle triple quotes
+					ind += 5;
+					while (ind < n && !(chars[ind] == '"' && chars[ind-1] == '"' && chars[ind-2] == '"'))
+						ind++;
+				} else {
+					// Handle double quotes
+					ind++;
+					while (ind < n && chars[ind] != '"' && chars[ind] != '\n')
+						ind++;
+				}
+				visitor.appendStyle(ind+1);
+				visitor.pop();					
+				break;
+			default:
+				// Check for token
+				boolean isLastCharacter = ind == n-1;
+//				boolean isWhitespace = Character.isWhitespace(c);
+				boolean isBreakingCharacter = !Character.isLetterOrDigit(c) && 
+						c != '.' &&
+						c != '_';
+				if (!isBreakingCharacter) {
+					buffer.append(c);
+				}
+				if (isBreakingCharacter)
+					handleToken(visitor, ind, buffer);
+				else if (isLastCharacter)
+					handleToken(visitor, n, buffer);
+				break;
+			}
+			// Increment counter
+			ind++;
 		}
 		
-        Matcher matcher = PATTERN.matcher(text);
-        int lastKwEnd = 0;
-        StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
-        while (matcher.find()) {
-            String styleClass =
-                    matcher.group("KEYWORD") != null ? "keyword" :
-                    matcher.group("PAREN") != null ? "paren" :
-                    matcher.group("BRACE") != null ? "brace" :
-                    matcher.group("BRACKET") != null ? "bracket" :
-                    matcher.group("SEMICOLON") != null ? "semicolon" :
-                    matcher.group("TRIPLEQUOTES") != null ? "string" :
-                    matcher.group("DOUBLEQUOTES") != null ? "string" :
-                    matcher.group("SINGLEQUOTES") != null ? "string" :
-                    matcher.group("COMMENT") != null ? "comment" :
-                    null; /* never happens */
-            assert styleClass != null;
-            spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
-            spansBuilder.add(Collections.singleton(styleClass), matcher.end() - matcher.start());
-            lastKwEnd = matcher.end();
-        }
-        spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
-        return spansBuilder.create();
+		var styles = visitor.buildStyles();
+		long endTime = System.currentTimeMillis();
+		logger.info("Style time: {} (length={})", endTime - startTime, n);
+		
+		return styles;
     }
+	
+	
+	private static void resetToken(StringBuffer buffer) {
+		buffer.setLength(0);
+	}
+	
+	
+	private static void handleToken(StyleSpanVisitor visitor, int ind, StringBuffer buffer) {
+		if (!buffer.isEmpty()) {
+			var s = buffer.toString();
+			int startInd = ind - s.length();
+			if (KEYWORDS.contains(s)) {
+				if (startInd > 0)
+					visitor.appendStyle(startInd);
+				visitor.push("keyword");
+				visitor.appendStyle(ind);
+				visitor.pop();
+			}
+			// Check for number
+			boolean isNumeric;
+			if (s.length() == 1) {
+				isNumeric = Character.isDigit(s.charAt(0));
+			} else {
+				isNumeric = NumberUtils.isCreatable(s);
+			}
+			if (isNumeric) {
+				if (startInd > 0)
+					visitor.appendStyle(startInd);
+				visitor.push("number");
+				visitor.appendStyle(ind);
+				visitor.pop();							
+			}
+		}
+		buffer.setLength(0);
+	}
+	
+	
+	private static void handleSingleCharacterStyle(StyleSpanVisitor visitor, int ind, String style) {
+		if (ind > 0)
+			visitor.appendStyle(ind);
+		visitor.push(style);
+		visitor.appendStyle(ind+1);
+		visitor.pop();
+	}
+	
 	
 	@Override
 	public StyleSpans<Collection<String>> computeConsoleStyles(final String text, boolean logConsole) {